refactor: 더미 데이터 수정
