+++ src/main/java/project/volunteer/DummyConfig.java
@@ -9,9 +9,8 @@ import project.volunteer.domain.participation.dao.ParticipantRepository;
 import project.volunteer.domain.recruitment.dao.RecruitmentRepository;
 import project.volunteer.domain.repeatPeriod.dao.RepeatPeriodRepository;
 import project.volunteer.domain.scheduleParticipation.dao.ScheduleParticipationRepository;
-import project.volunteer.domain.scheduleParticipation.domain.ScheduleParticipation;
 import project.volunteer.domain.sehedule.dao.ScheduleRepository;
-import project.volunteer.domain.storage.dao.StorageRepository;
+import project.volunteer.domain.image.dao.StorageRepository;
 import project.volunteer.domain.user.dao.UserRepository;
 
 @Configuration
